agent/requirements.txt
google-cloud-aiplatform>=1.60.0
google-auth>=2.29.0
streamlit>=1.35.0
PyGithub>=2.3.0
python-dotenv>=1.0.1
pyyaml>=6.0.1

